Card: Roberto Muratore, Adrian Vallejo (#879)
During the testing I found out the handling of pulled cards in card scripts (such as experimental) is not done properly, or not for all cards. To resolve this, I added new event handler into the pull function in player.js to handle the pull card every time a context is passed. Therefore the only responsibility card script has is to pass the context.