Refactor ability resolver order (#1082)
This change contains a big refactor change in ability resolver. The change is that the order of the cost and target resolution was swapped when abilities are resolved. Before, costs were fully resolved first and targets afterwards. After the swap, player first resolves targets and then costs. This will help with situations when you are shopping for attachments and cancel it, or you do not have proper target for the attachment. Also, thanks to the swap, cards such as Mount Moriah Cemetery can be implemented.